Refactor local
